+++ CHANGES.rst
@@ -7,6 +7,7 @@ Unreleased
 **Changed**
 
 - :meth:`.Redditor.moderated` will now objectify all data returned from the API.
+- :meth:`.Redditor.subreddit` will now objectify all data returned from the API.
 - The `wiki_edit` endpoint has been changed from `r/{subreddit}/api/wiki/edit/` to
   `r/{subreddit}/api/wiki/edit`.